(tests): Change the tests structure and add upload
